\name{cbPen}
\alias{cbPen}

\title{ ... }

\description{
...
}

\usage{
cbPen(cb, sp=-1, addSlag=NULL)
}

\arguments{
  \item{cb }{ ...}
  \item{sp }{ ...}
  \item{addSlag }{ ...}
}

\details{
...
}

\value{
...
}

\author{Antonio Gasparrini, \email{antonio.gasparrini@lshtm.ac.uk}}

\note{
... 
}

\seealso{
...

See \code{\link{dlnm-package}} for an introduction to the package and for links to package vignettes providing more detailed information.
}

\examples{
rnorm(1)
}

\keyword{utilities}
