\name{cr}
\alias{cr}

\title{ Generate a Basis Matrix ... }

\description{
...
}

\usage{
cr(x, df=10, knots=NULL, intercept=FALSE, fx= FALSE, S=NULL)
}

\arguments{
  \item{x }{ the predictor variable. Missing values are allowed.}
  \item{df }{ ...}
  \item{knots }{ ...}
  \item{intercept }{ ...}
  \item{fx }{ ...}
  \item{S }{ ...}
}

\details{
...
}

\value{
...
}

\author{Antonio Gasparrini, \email{antonio.gasparrini@lshtm.ac.uk}}

\note{
... 
}

\seealso{
...

See \code{\link{dlnm-package}} for an introduction to the package and for links to package vignettes providing more detailed information.
}

\examples{
rnorm(1)
}

\keyword{smooth}

