\name{crossplot}
\alias{crossplot}

\title{ Plot predicted effects for a DLNM }

\description{
Plot several graphs (3d, slices and overall effects) of predicted effects from distributed lag non-linear models (DLNM).
}

\usage{
crossplot(crosspred, type="3d", cumul=FALSE, ci="area", 
	var=NULL, lag=NULL, ylim=NULL, title=NULL, label="var")
}

\arguments{
  \item{crosspred }{ an object of class \code{"crosspred"}.}
  \item{type }{ type of plot. See Details below.}
  \item{cumul }{ logical. If \code{TRUE}, cumulative effects along lags are reported. Used only if \code{type="slices"}. See Details}
  \item{ci }{ type of confidence intervals representation, one of \code{"area"}, \code{"bars"} or \code{"lines"}.}
  \item{var, lag }{ vectors of predictor values and lags for which specific effects must be plotted. Used only if \code{type="slices"}.}
  \item{ylim }{ numeric vectors of length 2, giving the coordinates ranges for the response axis (y-axis for \code{type} equal to \code{"overall"} or \code{"slices"}, z-axis for \code{"3d"}).}
  \item{title }{ the main title on top of the plot.}
  \item{label }{ label for predictor to be inserted in the plot.}
}

\details{
Different plots can be obtained choosing the following values for the argument \code{type}:

\bold{\code{"3d"}}: a 3-D plot generated by calling the function \code{\link[graphics]{persp}}.

\bold{\code{"contour"}}: a contour/level plot generated by calling the function \code{\link[graphics]{filled.contour}}.

\bold{\code{"overall"}}: a plot of the overall effects (summed up all the single lag contributions).

\bold{\code{"slices"}}: a multiple plot of effects at specific values of predictor or lags, chosen by \code{var} and \code{lag}, respectively. Up to 4 plots for each dimension are allowed. Cumulative effects along lags are reported if \code{cumul=TRUE}: in this case, the same option must have been set to obtain the prediction saved in the \code{crosspred} object (see \code{\link{crosspred}}).

For a detailed illustration of the use of the functions, see:

\code{vignette("dlnmOverview")}
}

\references{ 
Armstrong, B. Models for the relationship between ambient temperature and daily mortality. \emph{Epidemiology}. 2006, \bold{17}(6):624-31.
}

\author{Antonio Gasparrini, \email{antonio.gasparrini@lshtm.ac.uk}}

\note{
All the effects are reported versus a reference value. For continuous functions, this is specified by the centering point defined in the object \code{crossbasis} (see \code{\link{crossbasis}}). Exponentiated effects are automatically returned if \code{model} has link equal to \code{log} or \code{logit}.

The values in \code{var} must match those specified in the object \code{crosspred} (see \code{\link{crosspred}}), while the values in \code{lag} must be included in the lag period specified by \code{\link{crossbasis}}.

This function creates plots with default settings (i.e. perspective in 3-D plot, colours etc.). Refer to the original estimates stored in the \code{crosspred} object in order to personalize the output with generic plot commands.
}

\seealso{ \code{\link{crossbasis}}, \code{\link{crosspred}}}

\examples{
# Example 3. See crossbasis and crosspred for other examples

### complex DLNM
### space of predictor: 5df quadratic spline for temperature
### space of predictor: linear effect for PM10
### lag function: 5df natural cubic spline for temperature up to lag30
### lag function: single strata at lag 0-1 for PM10

data(chicagoNMMAPS)
basis.pm <- crossbasis(chicagoNMMAPS$pm10,vartype="lin", lagtype="strata",
	cen=FALSE, maxlag=1)
basis.temp <- crossbasis(chicagoNMMAPS$temp, vartype="bs", vardf=5, vardegree=2,
	lagdf=5, cenvalue=21, maxlag=30)
summary(basis.pm)
summary(basis.temp)
model <- glm(death ~  basis.pm + basis.temp + ns(time, 7*14) + dow,
	family=quasipoisson(), chicagoNMMAPS)
pred.temp <- crosspred(basis.temp, model, at=-26:33)

crossplot(pred.temp, label="Temperature",
	title="3D graph of temperature effect")
crossplot(pred.temp, "contour", label="Temperature",
	title="Contour graph of temperature effect")
crossplot(pred.temp, "overall", label="Temperature",
	title="Overall effect of temperature over 30 days of lag")
crossplot(pred.temp, "slices", var=c(-20,0,27,33),
	lag=c(0,5,15,28), label="Temperature")

### See the vignette 'dlnmOverview' for a detailed explanation of this example
}

\keyword{hplot}
\keyword{smooth}
\keyword{ts}