% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm-package.R, R/fci.R, R/findrank.R, R/getcoef.R, R/getlink.R, R/getvcov.R, R/mkXpred.R, R/mkaddSlag.R, R/mkat.R, R/mkcen.R, R/mklag.R, R/seqlag.R
\name{internals}
\alias{dlnm-internal}
\alias{fci}
\alias{findrank}
\alias{getcoef}
\alias{getlink}
\alias{getvcov}
\alias{internals}
\alias{mkXpred}
\alias{mkaddSlag}
\alias{mkat}
\alias{mkcen}
\alias{mklag}
\alias{seqlag}
\title{Internal Functions for Package dlnm}
\usage{
fci(ci, x, high, low, ci.arg, plot.arg, noeff = NULL)

findrank(X)

getcoef(model, class)

getlink(model, class)

getvcov(model, class)

mkXpred(type, basis, at, predvar, predlag, cen)

mkaddSlag(addSlag, d)

mkat(at, from, to, by, range, lag, bylag)

mkcen(cen, type, basis, range)

mklag(lag)

seqlag(lag, by = 1)
}
\arguments{
\item{ci}{type of confidence intervals representation.}

\item{x}{the coordinates of the x axis.}

\item{high, low}{the coordinates of the y axis for the interval.}

\item{ci.arg}{list of arguments to draw the confidence intervals.}

\item{plot.arg}{list of arguments of the main plot.}

\item{noeff}{reference value of the null effect.}

\item{X}{a matrix.}

\item{model}{a regression model object.}

\item{class}{a regression model class.}

\item{type}{type of model from which predictions are needed. See \code{\link{crosspred}}.}

\item{basis}{type of basis object from which predictions are needed. See \code{\link{crosspred}}.}

\item{at}{either a numeric vector representing the values of a constant 
exposure throughout the lag period defined by \code{lag}, or a matrix of 
exposure histories over the same lag period used for estimation.}

\item{predvar}{vector or matrix of predictor values used for prediction}

\item{predlag}{vector or matrix of lag values used for prediction}

\item{cen}{logical or a numeric scalar. It specifies the centering value,
then used as a reference for predictions.}

\item{addSlag}{matrix or vector (or list of matrices and/or vectors)
defining additional penalties on the lag structure.}

\item{d}{numeric vector of length 2 providing the cross-basis dimensions.}

\item{from}{min of predictor values used for prediction.}

\item{to}{max of predictor values used for prediction.}

\item{by}{increment of the sequences of predictor used for prediction.}

\item{range}{range of values used for prediction.}

\item{lag}{either an integer scalar or vector of length 2, defining the
maximum lag or the lag range, respectively.}

\item{bylag}{increment of the sequences of lag values used for prediction.}

\item{at}{either a numeric vector representing the values of a constant 
exposure throughout the lag period defined by \code{lag}, or a matrix of 
exposure histories over the same lag period used for estimation.}

\item{cen}{logical or a numeric scalar. It specifies the centering value,
then used as a reference for predictions.}

\item{type}{type of model from which predictions are needed. See \code{\link{crosspred}}.}

\item{basis}{type of basis object from which predictions are needed. See \code{\link{crosspred}}.}

\item{range}{range of values used for prediction.}

\item{lag}{either an integer scalar or vector of length 2, defining the
maximum lag or the lag range, respectively.}

\item{lag}{either an integer scalar or vector of length 2, defining the
maximum lag or the lag range, respectively.}
}
\description{
These functions are called internally by other functions and are not meant
to be directly run by the users.
}
\section{Functions}{
\itemize{
\item \code{fci}: \code{fci} provides different options for representing
confidence intervals, and it is called internally in plotting functions.

\item \code{findrank}: \code{findrank} returns the rank of a matrix.

\item \code{getcoef}: extracts coefficients depending on the class of the model, 
and returns a message error if the process fails. Used internally in \code{\link{crosspred}} and
\code{\link{crossreduce}}.

\item \code{getlink}: extracts link function depending on the class of the model, 
and returns a message error if the process fails. Used internally in \code{\link{crosspred}} and
\code{\link{crossreduce}}.

\item \code{getvcov}: extracts coefficient covariance depending on the class of the model, 
and returns a message error if the process fails. Used internally in \code{\link{crosspred}} and
\code{\link{crossreduce}}.

\item \code{mkXpred}: used internally in \code{\link{crosspred}} to
define the basis or cross-basis matrix for computing predictions.

\item \code{mkaddSlag}: returns a list of penalty matrices defining
additional penalties on the lag structure.

\item \code{mkat}: used internally in \code{\link{crosspred}} to define
the values used for predictions.

\item \code{mkcen}: used internally in \code{\link{crosspred}} to
define the centering value for computing predictions.

\item \code{mklag}: used internally in several other functions to
check/define the vector of length 2 representing the lag interval.

\item \code{seqlag}: used internally in several other functions to
create the vector with the sequence of lags given the range provided.
}}
\note{
These and other undocumented internal functions are not exported in
the namespace of the package \pkg{dlnm}. The user can visualize them through
the use of the triple colon operator '\code{:::}' or through the function
\code{\link{getAnywhere}}.
}
\author{
Antonio Gasparrini <\email{antonio.gasparrini@lshtm.ac.uk}>
}
\seealso{
See \code{\link{dlnm-package}} for an introduction to the package
and for links to package vignettes providing more detailed information.
}
\keyword{internal}

