\name{smooth.construct.cb.smooth.spec}
\alias{smooth.construct.cb.smooth.spec}
\alias{Predict.matrix.cb.smooth}

\title{ ... }

\description{
...
}

\usage{
smooth.construct.cb.smooth.spec(object, data, knots)

Predict.matrix.cb.smooth(object, data)
}

\arguments{
  \item{object }{ ...}
  \item{data }{ ...}
  \item{knots }{ ...}
}

\details{
...
}

\value{
...
}

\author{Antonio Gasparrini, \email{antonio.gasparrini@lshtm.ac.uk}}

\note{
... 
}

\seealso{
...

See \code{\link{dlnm-package}} for an introduction to the package and for links to package vignettes providing more detailed information.
}

\examples{
rnorm(1)
}

\keyword{smooth} \keyword{models} \keyword{regression}
