% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlnm-package.R
\docType{package}
\name{dlnm-package}
\alias{dlnm}
\alias{dlnm-package}
\title{Distributed Lag Non-linear Models (DLNM)}
\description{
The package \pkg{dlnm} contains functions to specify and run distributed lag
linear and non-linear models. These functions are used to build basis and
cross-basis matrices and then to predict and plot the results for a fitted
model.
}
\section{Modelling framework}{
 Distributed lag non-linear models (DLNM)
represent a modelling framework to describe simultaneously non-linear and
delayed dependencies, termed as \emph{exposure-lag-response associations}.
The methodology of DLNMs was originally developed for time series data, and
has been recently extended to other study designs and data structures,
compatible with cohort, case-control or longitudinal studies, amongst
others. A thorough methodological overview is given in the references and
the package vignettes detailed below.

The modelling framework is based on the definition of a \emph{cross-basis},
a bi-dimensional space of functions specifying the dependency along the
space of the predictor and along lags. The cross-basis functions are built
combining the basis functions for the two dimensions, produced by applying
existing or user-defined functions such as splines, polynomials, linear
threshold or indicators. The DLNM family includes simple distributed lag
models (DLM) as a special case.

The application of DLNMs requires the availability of predictor values at
equally-spaced time points. In the original development in time series
analysis, these are represented by the ordered series of observations. More
generally, the data can be stored in a matrix of \emph{exposure histories},
where each row represents the lagged values of the predictor for each
observation.

The cross-basis matrix of transformed variables is included in the model
formula of a regression model to estimate the associated parameters. The
estimation can be carried out with the default regression functions, such as
\code{\link{lm}}, \code{\link{glm}}, \code{gam} (package \pkg{mgcv}),
\code{clogit} and \code{coxph} (package \pkg{survival}), \code{lme} (package
\pkg{nlme}), \code{lmer} and \code{glmer} (package \pkg{lme4}). Estimates
are then extracted to obtain predictions and graphical representations which
facilitate the interpretation of the results.
}
\author{
Antonio Gasparrini and Ben Armstrong, with contributions from Fabian
Scheipl

Maintainer: Antonio Gasparrini <\email{antonio.gasparrini@lshtm.ac.uk}>
}
\references{
Gasparrini A. Distributed lag linear and non-linear models in R:
the package dlnm. \emph{Journal of Statistical Software}. 2011;
\bold{43}(8):1-20, freely available at
\url{http://www.ag-myresearch.com/jss2011}.

Gasparrini A. Modeling exposure-lag-response associations with distributed
lag non-linear models. \emph{Statistics in Medicine}. 2014;
\bold{33}(5):881-899, freely available at
\url{http://www.ag-myresearch.com/statmed2014}.

Gasparrini A., Armstrong, B.,Kenward M. G. Distributed lag non-linear
models. \emph{Statistics in Medicine}. 2010; \bold{29}(21):2224-2234,
freely available at \url{http://www.ag-myresearch.com/statmed2010}.

Gasparrini A., Armstrong, B., Kenward M. G. Reducing and meta-analyzing
estimates from distributed lag non-linear models.\emph{BMC Medical Research
Methodology}. 2013; \bold{13}(1):1, freely available at
\url{http://www.ag-myresearch.com/bmcmrm2013}.

Armstrong, B. Models for the relationship between ambient temperature and
daily mortality. \emph{Epidemiology}. 2006, \bold{17}(6):624-31, available
at \url{http://www.ncbi.nlm.nih.gov/pubmed/17028505}.
}
\seealso{
\code{\link{onebasis}} to generate simple basis matrices.
\code{\link{crossbasis}} to generate cross-basis matrices.
\code{\link{crosspred}} to obtain predictions after model fitting.
\code{\link{crossreduce}} to reduce the fit to one dimension. The methods
\code{\link{plot.crosspred}} and \code{\link{plot.crossreduce}} to plot
several type of graphs.

Type \code{'vignette(dlnmOverview)'} for a detailed description.
}
\keyword{package}

